% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_functional.R
\name{bwNeweyWest1987}
\alias{bwNeweyWest1987}
\title{Bandwidth as in Newey and West 1987}
\usage{
bwNeweyWest1987(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{...}
}
\value{
returns the selected bandwidth parameter
}
\description{
Used in estimate functional to describe bandwidth selection as proposed in Newey and West (1987).
Applies bandwidth that includes lags accroding to $m(T)=T^{1/5}$.
}
