% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_functional.R
\name{expectiles}
\alias{expectiles}
\title{Identification function for state-dependent expectiles}
\usage{
expectiles(x, y, stateVariable, theta, model, ...)
}
\arguments{
\item{x}{forecast}

\item{y}{realization}

\item{stateVariable}{state variable}

\item{theta}{model parameter to be estimated}

\item{model}{model function}

\item{...}{...}
}
\description{
Identification function for state-dependent expectiles
}
\examples{
## Estimate expectile level for constant specification model with estimate.functional

res <- estimate.functional(Y=GDP$observation, X=GDP$forecast,
model=constant,
instruments="const",
iden.fct = expectiles)
summary(res)
plot(res)
}
\seealso{
Other identification functions: \code{\link{quantiles}}
}
\concept{identification functions}
