% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_functional.R
\name{lag}
\alias{lag}
\title{Lagging variables for use in estimate functional}
\usage{
lag(vector, lag = 1)
}
\arguments{
\item{vector}{vector to be lagged}

\item{lag}{number of lags}
}
\value{
lagged vector of same length with NAs at beginning
}
\description{
Lagging variables for use in estimate functional
}
\examples{
#lag example vector by one lag
lag(c(1,2,3))

#lag example vector by two lags
lag(c(1,2,3,4),lag=2)
}
