% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_functional.R
\name{logistic}
\alias{logistic}
\title{logistic specification model}
\usage{
logistic(stateVariable, theta)
}
\arguments{
\item{stateVariable}{state variable}

\item{theta}{parameter}
}
\value{
numeric level
}
\description{
logistic specification model
}
