% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointFore.R
\docType{package}
\name{PointFore}
\alias{PointFore}
\alias{PointFore-package}
\title{PointFore: A package for computating the notorious bar statistic.}
\description{
Estimate specification models for the state-dependent level of an optimal quantile/expectile forecast.
Wald Tests and the test of overidentifying restrictions are implemented. Plotting of the estimated specification model is possible.
The package contains daily accumulated precipitation at London, UK from the high-resolution model of the
European Centre for Medium-Range Weather Forecasts (ECMWF, https://www.ecmwf.int/).
Based on "Interpretation of Point Forecasts" by Patrick Schmidt, Matthias Katzfuss, and Tilmann Gneiting.
}
\section{PointFore functions}{

The main function is estimate.functional(). It returns an object which can be analyzed with the plot() and summary() function.
}

