% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precipitation}
\alias{precipitation}
\title{daily accumulated precipitation (in millimeter) at London, UK and 24-hours-ahead forecasts of the ECMWF (1969-2011)}
\format{A data frame with 2192 rows and 2 variables:
\itemize{
  \item Y: daily accumulated precipitation in millimeter at London, UK (0 -- 45)
  \item X: according point forecast in millimeter issued one day ahead (0 -- 40)
}}
\usage{
precipitation
}
\description{
24-hour ahead forecasts of daily accumulated precipitation at London,
UK from the high-resolution model of the
European Centre for Medium-Range Weather Forecasts (ECMWF). The data contains observations from 2012 to 2016.
}
\details{
We thank the ECMWF for their support. Further details can be found on \url{https://www.ecmwf.int/}.
}
\keyword{datasets}
