% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_functional.R
\name{probit}
\alias{probit}
\title{propit specification model}
\usage{
probit(stateVariable, theta)
}
\arguments{
\item{stateVariable}{state variable}

\item{theta}{parameter}
}
\value{
numeric level
}
\description{
propit specification model
}
