% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_functional.R
\name{constant}
\alias{constant}
\title{Constant specification model}
\usage{
constant(stateVariable, theta, ...)
}
\arguments{
\item{stateVariable}{state variable}

\item{theta}{parameter}

\item{...}{...}
}
\value{
numeric level
}
\description{
Constant specification model
}
