% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_functional.R
\name{estimate.functional}
\alias{estimate.functional}
\title{Estimate Functional}
\usage{
estimate.functional(iden.fct = quantiles, model = constant, theta0 = 0.5,
  Y, X, stateVariable = NULL, other_data = NULL, instruments = c("X",
  "lag(Y)"), ...)
}
\arguments{
\item{iden.fct}{identification function. Standard choice is quantiles. Alternative is expectiles.}

\item{model}{specification model}

\item{theta0}{starting value for optimization}

\item{Y}{realized values}

\item{X}{forecasts}

\item{stateVariable}{state variable(s)}

\item{other_data}{optional for construction of instruments}

\item{instruments}{instruments (list of character describing instruments or matrix of actual intsruments)}

\item{...}{other parameters for gmm function (see ?gmm)}
}
\value{
Object of type pointfore
}
\description{
Estimates the parameter in a specification model for state-dependent quantile or expectile forecasts.
}
\examples{
estimate.functional(Y=precipitation[,1],X=precipitation[,2])
estimate.functional(Y=precipitation[,1],X=precipitation[,2],
instruments=c("X","lag(Y)"))
estimate.functional(Y=precipitation[,1],X=precipitation[,2],
other_data = data.frame(Z=precipitation[,2]),
instruments=c("Z","lag(Y)"))
estimate.functional(Y=precipitation[,1],X=precipitation[,2],
model=logistic,theta0=c(0,0),stateVariable = precipitation[,2])
}
