% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDP}
\alias{GDP}
\title{real GDP realized values and one quarter ahead Greenbook forecasts (1969-2011)}
\format{A data frame with 172 rows and 2 variables:
\itemize{
  \item first: realized GDP growth rate in percentage measured at first vintage (-4.7--8.5)
  \item gRGDPF1: according point forecast issued one quarter before (-10.4--11.2)
}}
\source{
https://www.philadelphiafed.org/research-and-data/real-time-center/greenbook-data/philadelphia-data-set
}
\usage{
GDP
}
\description{
A dataset containing real GDP growth rate in the United States and according one quarter ahead point forecasts from Federal Reserve's Greenbook.
The forecasts were selected to be closest to the middle of the respective quarter among the published Greenbook forecasts of one quarter.
}
\keyword{datasets}
