% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_functional.R
\name{logistic_linear}
\alias{logistic_linear}
\title{logistic-linear specification model}
\usage{
logistic_linear(stateVariable, theta, ...)
}
\arguments{
\item{stateVariable}{state variable}

\item{theta}{parameter}
}
\value{
numeric level
}
\description{
logistic-linear specification model
}
