% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pointfore}
\alias{plot.pointfore}
\title{Plots object of class "pointfore"}
\usage{
\method{plot}{pointfore}(x, conf.levels = c(0.6, 0.9), pdf = TRUE,
  hline = NULL, adjust.factor = 1, limits = NULL, ...)
}
\arguments{
\item{x}{object of class "pointfore"}

\item{conf.levels}{one or two confidence levels for pointwise confidence intervals}

\item{pdf}{logic if pdf estimate should be plotted}

\item{hline}{if TRUE plots horizontal line at 0.5. if numeric plot horizontal line at value.}

\item{adjust.factor}{adjust factor for estimating pdf (controls smoothness)}

\item{limits}{2-dimensional vector defining range of x-axis}

\item{...}{other parameters}
}
\value{
plot
}
\description{
Plots object of class "pointfore"
}
\examples{
res <- estimate.functional(Y=GDP$observation,X=GDP$forecast,
model=logistic,theta0=c(0,0),
stateVariable = lag(GDP$observation))
plot(res)
}
