% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_functional.R
\name{probit_linear}
\alias{probit_linear}
\title{linear specification model with probit link}
\usage{
probit_linear(stateVariable, theta, ...)
}
\arguments{
\item{stateVariable}{state variable}

\item{theta}{parameter}

\item{...}{other parameters}
}
\value{
numeric level
}
\description{
Specification models are used to denote the quantile or expectile level (depending on the identifciation function).
The linear probit specification model depends linear on the state variable with a probit link function.
}
\examples{
# plot linear probit specification model with constant quantile/expectile level
plot(function(x) probit_linear(x,theta=c(0,0)), xlim=c(-1,1))


# plot linear probit specification model with state-dependent quantile/expectile level
plot(function(x) probit_linear(x,theta=c(0,5)), xlim=c(-1,1))
}
