% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_functional.R
\name{probit_linear}
\alias{probit_linear}
\title{linear specification model with probit link}
\usage{
probit_linear(stateVariable, theta)
}
\arguments{
\item{stateVariable}{state variable}

\item{theta}{parameter}
}
\value{
numeric level
}
\description{
linear specification model with probit link
}
