% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_functional.R
\name{probit_spline2}
\alias{probit_spline2}
\title{quadratic spline specification model with probit link}
\usage{
probit_spline2(stateVariable, theta, ...)
}
\arguments{
\item{stateVariable}{state variable}

\item{theta}{parameter}

\item{...}{...}
}
\value{
numeric level
}
\description{
quadratic spline specification model with probit link
}
\examples{
# plot example of cubic spline specification model with state-dependent quantile/expectile level
plot(function(x) probit_spline3(x,theta=c(0,2,1,-1)), xlim=c(-1,1))
}
