% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_functional.R
\name{probit_spline3}
\alias{probit_spline3}
\title{cubic spline specification model with probit link}
\usage{
probit_spline3(stateVariable, theta, ...)
}
\arguments{
\item{stateVariable}{state variable}

\item{theta}{parameter}

\item{...}{...}
}
\value{
numeric level
}
\description{
cubic spline specification model with probit link
}
