% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_functional.R
\name{quantiles}
\alias{quantiles}
\title{Identification function for state-dependent quantiles}
\usage{
quantiles(x, y, stateVariable, theta, model, ...)
}
\arguments{
\item{x}{forecast}

\item{y}{realization}

\item{stateVariable}{state variable}

\item{theta}{model parameter to be estimated}

\item{model}{model function}
}
\description{
Identification function for state-dependent quantiles
}
